move loading emoji outside the worker file in case workers don't exist
