+++ app/javascript/mastodon/features/emoji/loader.ts
@@ -1,0 +1,75 @@
+import type { FlatCompactEmoji } from 'emojibase';
+
+import type { ApiCustomEmojiJSON } from '@/mastodon/api_types/custom_emoji';
+import { isDevelopment } from '@/mastodon/utils/environment';
+
+import {
+  putEmojiData,
+  putCustomEmojiData,
+  loadLatestEtag,
+  putLatestEtag,
+} from './database';
+import { toSupportedLocale, toSupportedLocaleOrCustom } from './locale';
+import type { LocaleOrCustom } from './locale';
+
+export async function importEmojiData(localeString: string) {
+  const locale = toSupportedLocale(localeString);
+  const emojis = await fetchAndCheckEtag<FlatCompactEmoji[]>(locale);
+  if (!emojis) {
+    return;
+  }
+  await putEmojiData(emojis, locale);
+}
+
+export async function importCustomEmojiData() {
+  const emojis = await fetchAndCheckEtag<ApiCustomEmojiJSON[]>('custom');
+  if (!emojis) {
+    return;
+  }
+  await putCustomEmojiData(emojis);
+}
+
+async function fetchAndCheckEtag<ResultType extends object[]>(
+  localeOrCustom: LocaleOrCustom,
+): Promise<ResultType | null> {
+  const locale = toSupportedLocaleOrCustom(localeOrCustom);
+  const oldEtag = await loadLatestEtag(locale);
+
+  let uri: string;
+  if (locale === 'custom') {
+    uri = '/api/v1/custom_emojis';
+  } else {
+    uri = `/packs${isDevelopment() ? '-dev' : ''}/emoji/${locale}.json`;
+  }
+
+  const response = await fetch(uri, {
+    headers: {
+      'Content-Type': 'application/json',
+      'If-None-Match': oldEtag ?? '', // Send the old ETag to check for modifications
+    },
+  });
+  // If not modified, return null
+  if (response.status === 304) {
+    return null;
+  }
+  if (!response.ok) {
+    throw new Error(
+      `Failed to fetch emoji data for ${localeOrCustom}: ${response.statusText}`,
+    );
+  }
+
+  const data = (await response.json()) as ResultType;
+  if (!Array.isArray(data)) {
+    throw new Error(
+      `Unexpected data format for ${localeOrCustom}: expected an array`,
+    );
+  }
+
+  // Store the ETag for future requests
+  const etag = response.headers.get('ETag');
+  if (etag) {
+    await putLatestEtag(etag, localeOrCustom);
+  }
+
+  return data;
+}
